+++ catalog/app/@types/index.d.ts
@@ -1,0 +1,8 @@
+declare module '*.png' {
+  const value: string
+  export default value
+}
+
+declare module 'intl/locale-data/jsonp/*'
+
+type $TSFixMe = any