+++ catalog/app/@types/mui.d.ts
@@ -1,0 +1,10 @@
+import * as React from 'react'
+import '@material-ui/core'
+
+declare module '@material-ui/core' {
+  // workaround for MUI@4 not having ref in BoxProps
+  // https://github.com/mui-org/material-ui/issues/17010
+  interface BoxProps {
+    ref?: React.Ref<any>
+  }
+}